requirements.txt
asgiref==3.8.1
Django==5.2.1
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==25.0
python-dotenv==1.1.0
sqlparse==0.5.3
tzdata==2025.2
